Move a few links to ‘backward’
If a Link is just an alternate name, it’s not needed for zone.tab
and ‘backward’ is a better home for it anyway.  This helps improve
link checking.
* Makefile (check_links): Use new backcheck option.
* asia (Europe/Nicosia):
* etcetera (Etc/Universal, Etc/Zulu, Etc/Greenwich, Etc/GMT-0)
(Etc/GMT+0, Etc/GMT0):
* europe (Asia/Istanbul):
Move Link to ‘backward’.
* checklinks.awk: Add backcheck option, to check that
non-‘backward’ Links are next to the Zones they link to, and that
a Link is in a continental file if and only if it is in zone.tab.
Add a special case for GMT.
* etcetera (GMT): In vanguard form, now a Zone, not a Link,
since it’s the file the leap second code depends on.
(Etc/GMT): In vanguard form, now a link in ‘backward’ instead of a
Zone here.
* ziguard.awk: Support the vanguard form.