Fix usage of __has_include
Problem with GCC 4.8.3 reported by Igor Ivanov in:
https://mm.icann.org/pipermail/tz/2022-October/032129.html
* private.h (true, false, bool) [__STDC_VERSION__ < 202311]:
Define early so that we needn’t worry about using these macros
before they’re defined.
(HAS_INCLUDE): Remove.  This macro was misguided, as C23 says the
behavior is undefined if a macro body expands to
__has_include(...).  All uses of HAS_INCLUDE replaced by direct
invocations of __has_include; or if the old code used
__STDC_VERSION__ just go back to that as it still works for
__STDC_VERSION__ == 202311.  This patch also works around the
problem reported for GCC 4.8.3.