Use ‘restrict’ more consistently with standards
Previously, I assumed that if the declaration of a function F
declared an argument via ‘T *restrict ARG’ then the compiler could
take advantage of the ‘restrict’ when compiling F’s definition.
However, apparently this is not the case: a top-level ‘restrict’
like that has no effect semantically unless it’s present in F’s
definition.  Fix this by using ‘restrict’ consistently in both
declarations and definitions of functions where POSIX or the
NetBSD API specifies ‘restrict’.  (As an exception, do not use
‘restrict’ in tzfree’s argument as this appears to be a typo in
NetBSD.)  Do not bother using ‘restrict’ in internal functions, as
this change is for consistency not for optimization.  This change
was prompted by a bug report for compiling on MS-Windows; see
https://mm.icann.org/pipermail/tz/2022-December/032451.html
Although the change does not fix the MS-Windows situation, the
idea is to fix the MS-Windows issue another way.
* asctime.c (asctime_r):
* localtime.c (localtime_rz, localtime_r, gmtime_r, mktime_z):
* private.h (asctime) [HAVE_INCOMPATIBLE_CTIME_R]:
* strftime.c (strftime_l, strftime):
Use ‘restrict’ consistently when POSIX or tzcode’s doc say so.