Support leap-seconds.list tombstone
This change looks forward to the possibility of leap seconds being
discontinued by 2035.  If and when that happens, the expiration
date on the leap-seconds.list file is likely to stop being
maintained, as there will be no need to issue new versions of
leap-seconds.list every six months.  Although the details have not
been decided, for now assume that the last edition of
leap-seconds.list will simply lack an expiration-date line.
* leapseconds.awk: If the input lacks an expiration line like
‘#@ 3896899200’, do not treat it as if ‘#@ 0’ were present, as
that would mean an expiration date of 1900-01-01.  Instead, do not
output an expiration date; just output a comment saying that there
is no expiration date.  Similarly for update dates and input
lacking lines like ’#$ 3676924800’.