Bring doc up-to-date with draft POSIX
* Makefile, date.1, newctime.3, newtzset.3, theory.html:
Don’t say settings like TZ='EST5EDT' are “nonstandard and obsolete”,
as draft POSIX says they use implementation-defined DST rules.
Clarify and strengthen the wording of the description of the
now-obsolete POSIXRULES macro and posixrules file, the latter of
which is still commonly installed even though it does not work.
Improve doc for TZDEFRULESTRING.  Give Morocco instead of
Iran as an example of unusual DST, as Iran no longer observes DST.