Update title and description of Tom Scott video
The video was originally published 2019-02-18 with the title "Why
Denmark Is .07 Seconds Behind The World".  The value of UT1-UTC was
programmatically kept updated through 2022: (at 4:45) "I've written some
code that will keep checking the clocks and automatically update the
title of this video as we go.  At least until the code breaks and I
can't be bothered to fix it anymore" -- until it was updated in June
2023 to reflect the past tense.

See:
https://mm.icann.org/pipermail/tz/2020-October/029364.html
https://mm.icann.org/pipermail/tz/2023-February/032585.html