Fix localtime bug with unusual TZif data
To reproduce the bug, use the following file foo.zi:
	Rule R 2073 max  - Mar Sat<=30 2:00 1:00 S
	Rule R 2073 max  - Oct Sat<=30 2:00 0    -
	Zone Z 2:00 R EE%sT
and run ‘zic foo.zi; zdump -i -c 2072,2073 Z’.  Without the fix,
this incorrectly outputs the following two lines:
	2072-03-26	03	+03	EEST	1
	2072-10-29	01	+02	EET
even though there are no transitions in 2072.
* localtime.c (tzloadbody): Compute sp->goahead from ts->goahead.
Do not guess goahead and goback from the timestamps, as that was
just a guess that predates having TZ strings in the TZif file, and
the guess is no longer needed and sometimes guesses wrong.
(typesequiv): Remove; no longer used.
(tzparse): Last arg is now pointer to const.  This is just for
clarity; it doesn’t affect behavior.