Fix zic bug with Palestine after 2075
The bug can be observed when processing the following .zi data,
adapted from the current ‘asia’ file:
	Rule Palestine 2075 max  - Mar Sat<=30 2:00 1:00 S
	Rule Palestine 2075 max  - Oct Sat<=30 2:00    0 -
	Rule Palestine 2076 only - Jul      25 2:00    0 -
	Rule Palestine 2076 only - Sep       5 2:00 1:00 S
	Zone Asia/Gaza 2:00 -         EET   2012
	               2:00 Palestine EE%sT
Without the fix, zic generates an incorrect TZif file, in which
the special-case 2076 transitions are omitted.  This causes ‘zdump
-ic 2076,2077 Asia/Gaza’ to mistakenly omit the lines:
	2076-07-25	01	+02	EET
	2076-09-05	03	+03	EEST	1
* zic.c (outzone): Redo algorithm to work even when the effect of
a Rule that never ends (TO="max") is interspersed with the effect
of a one-shot rule (TO="only").