make Makefile more compatible with POSIX
This was initially motivated by a desire to have the Makefile be
friendlier to new builders, so that the Makefile does not use CC in
an unusual way, and so that builders can run commands like
‘make CC='gcc -std=gnu11'’ without messing up TZDIR.
While I was at it, I fixed other ways that the Makefile is
inconsistent with modern POSIX.
* Makefile (CFLAGS, LDFLAGS, AR, CC): Do not define.  Instead,
rely on system-supplied defaults, which are specified by POSIX.
(SHELL): Do not define or use, as POSIX prohibits that.
(cc): Remove.  All uses changed to use CC, which is where
POSIX expects the C compiler to be.
(DTZDIR): New macro, so that TZDIR is easily specified
regardless of CC or CFLAGS.
(zdump, zic, tzdir, check_public, $(TIME_T_ALTERNATIVES))
(typecheck_long_long, typecheck_unsigned):
Use TZDIR and pass it along.
(libtz.a): Use ARFLAGS rather than open-coding the default.
* NEWS: Mention these changes.