For strftime %z, use tm_gmtoff if available
Problem and draft patch reported by Dag-Erling Smørgrav in:
https://mm.icann.org/pipermail/tz/2024-January/033488.html
* NEWS: Mention this.
* localtime.c (EXTERN_TIMEOFF): Default to static.
(timeoff): Use EXTERN_TIMEOFF to declare; this is easier to read.
* private.h (EXTERN_TIMEOFF): New macro.
(timeoff): Define to tz_private_timeoff if it’s not public but
strftime needs it anyway.  Declare if we define it as extern.
* strftime.c (_fmt): If TM_GMTOFF is available, use timeoff
rather than mktime to determine %z.  Do not worry about UNINIT_TRAP
since we are now using a more-generous reading of POSIX.