Conform to RFC 8536 section 3.2 for default type
Remove a heuristic for reading TZif files that were generated by
tzcode 2018e and earlier.  The heuristic was designed to work
around problems with these TZif files, which caused localtime in
some cases to misbehave for early timestamps before the first
transition.  Although helpful for reading nonstandard old data,
the heuristic causes localtime to fail to conform to Internet RFC
8536 section 3.2, which says “Local time for timestamps before the
first transition is specified by the first time type (time type	0).”
Nowadays it is safe to assume tzcode 2018f or later, so there is
no need for this nonstandard heuristic.
* NEWS: Mention this.
* localtime.c (struct state): Remove defaulttype member.  All uses
removed, and replaced by 0.