Remove now-dead strftime code
We didn’t update the manual and we left in some now-dead code in
commit 0e8f0b06ac2f65d0c54a91121da09c8662cd88e6 dated 2022-06-06
18:57:23 -0700, which implemented Elliott Hughes’s suggestion in
https://mm.icann.org/pipermail/tz/2022-June/031485.html
to have strftime %s format (time_t)-1 when a timestamp was out of
range for time_t.  Problem reported by Tom Lane in:
https://lists.iana.org/hyperkitty/list/tz@iana.org/thread/TC4GEIRORDGM7F2UQEJCZF6BFGXUNKUV/
* newstrftime.3: Improve documentation of %s’s glitches.
Don’t say that our strftime can fail with EOVERFLOW, but say that
it’s possible in other implementations (and they can also fail
with EINVAL).
* strftime.c (strftime): Remove code that assumed that _fmt could
return a null pointer.