Remove dependency of asctime on strftime
Also, remove calls to strlen and strcpy.
* asctime.c (ASCTIME_FMT, ASCTIME_FMT_B):
Remove; the strings are now merely literals in asctime_r.
(SNPRINTF, SNPRINTFBUF): New macros.
(asctime_r): Do not call strftime %Y, as this is dicey when
tm_year is not a value that could be yielded by localtime.
Instead, use snprintf, or sprintf on ancient hosts lacking snprintf.
Also, rely on snprintf/sprintf's return value instead of calling
strlen, as it's safe to assume C89 b6 now.  If snprintf is available,
output to the result buffer directly instead of to the stack and then
copying.  This saves one call to strftime, two calls to strlen, and
one call to strcpy.
* private.h (HAVE_SNPRINTF): Move here ...
* zdump.c: ... from here.