Fix bug with -d RELATIVE -t ABSOLUTE
Problem reported by Evgeniy Gorbanev in:
https://lists.iana.org/hyperkitty/list/tz@iana.org/thread/Y4IQO6VNGAALRTS2TEACPHZBUMTIPHEZ/
Fix the main bug here by not trying to create a symlink
with contents that won’t work because they’re relative.
This also fixes the read buffer underflow.
While we’re at it, fix some bogus diagnostics issued when
a file name is absolute.
* NEWS: Mention this.
* zic.c (close_file): If NAME is absolute, don’t output any
directory that it’s relative to.
(directory_end_in_slash): New static var.
(main): Set it after changing directory, in case directory is empty.
(diagdir, diagslash): New functions.
(open_outfile, rename_dest, dolink): Use them in diagnostics to
handle the unusual case where an absolute file name is used.
(relname): Return a null pointer if LINKNAME is absolute but
DIRECTORY is not, because the computed symlink wouldn’t work
in that case.
(dolink): If relname returns a null pointer, don’t try
to create a symlink; just make a copy.  When warning about the
copy, say that the symlink is not obvious.