Better fix for U+2217 ASTERISK OPERATOR glitch
* newctime.3, newtzset.3: Use plain * instead of \(**.
* workman.sh: Remove no-longer needed workaround for the U+2217
ASTERISK OPERATOR glitch.  This was not a glitch in groff; it was
a glitch in TZDB’s man pages dating back to 1986, presumably
because back then \(** generated nicer-looking output on the C/A/T
phototypesetter whereas ordinary * was raised.  Nowadays It’s
better to use plain * as it looks like an ASCII * and it works if
you cut and past it into C code.