Prefer mempcpy to strcpy from non-constant
* localtime.c (tzloadbody, tzset_unlocked): When the source string
is not known to be constant, prefer mempcpy (and terminating the
destination by hand instead of trusting the source) to strcpy.
This avoids some undefined behavior if another thread modifies the
data while we access it.  With typical optimization the resulting
machine code will not call strcpy because all uses of strcpy are
optimized away.