Omit no-op transitions when Rule+Zone cancel
Fix zic so that it does not generate a no-op transition for
TZDB 2025c’s Asia/Tbilisi at POSIX time 859662000, i.e.,
at 1997-03-29 19:00 UTC or 1997-03-30 00:00 local time.
The no-op transition was generated because there is a Zone line
"4:00 1:00 %z 1997 Mar lastSun" followed by "4:00 E-EurAsia %z ..."
and the relevant Rule as of 1997 Mar lastSun is
"Rule E-EurAsia 1981 max - Mar lastSun 0:00 1:00 -".
Because the first Zone line’s UNTIL and the Rule’s time are the same,
zic coalesced the two into a no-op transition where the
timestamps before and after are both +05 with DST.
This change also fixes a bug with the following test:
  Rule NOS 1970 2037 - Jan 27 23:43:01 2:0 SDST
  Rule NOS 1970 2037 - Jan 28 0:51:01 0 S
  Zone OS/NOS +7:0:0 NOS NO%s
In this unrealistic test, the fall-back occurs during the
spring-forward gap so there should be no TZif transitions.
Without the fix zic incorrectly outputs a 2-hour spring-forward
transition from 1970-01-27 22:51:01 +07 to 1970-01-28 00:51:01 +09,
and then outputs 66 no-op transitions every year until 2037-01-28
00:51:01 +09, and then the non-matching TZ string "NOS-7".
This test case was reported by Renchunhui in:
https://lists.iana.org/hyperkitty/list/tz@iana.org/message/TEUFHPYKORVAXWQV5FTNAOAQOXEMQDWZ/
* NEWS: Mention this.
* zic.c (writezone): When removing an unwanted transition
due to a Rule line applying at the same time as a Zone’s UNTIL time,
do not turn the preceding transition into a no-op.