Port to POSIX.1-2001 fflush
Problem on NetBSD reported by Christos Zoulas in:
https://lists.iana.org/hyperkitty/list/tz@iana.org/thread/I4CHNIVPXFXVISCGCJMW4OHD4N7O3PTQ/
* NEWS: Mention this.
* zic.c (close_file): Call fflush only if tempname is nonnull.
As Jonathan Wakely pointed out in the above-cited thread,
the requirement that fflush succeed on read-only streams
was added in POSIX.1-2008, and NetBSD’s fflush conforms only
to POSIX.1-2001 in this department.  Ironically, the fflush
was put in only to port to non-POSIX platforms where you can’t
write to a file after chmodding it to be read-only.